Remove use of Go routines and WaitGroup
The functions that were being run as Go routines are short and cpu
bound. Not running them as Go routines simplifies the code, and is
faster.

The wall clock time for the execution of the code in parsers.GetArtist
(_after_ requests.FetchArtist) is reduced when not using Go routines.
It dropped from around 20μS to about 5μS.